AuddForwarder/requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
python-dotenv==1.2.2
requests==2.33.1
python-multipart==0.0.26
openai==2.32.0
